refactor: move some configuration-accessor logics in command
`configuration-accessor` is bloated. Let's move the functionalities related to spawning a recorded process in a dedicated component. The idea of `configuration-accessor` was to control the access to `configuration` objects. But I don't such control is actually needed. So in the future we might get rid of `configuration-accessor` altogether.
