+++ components/command/node/.env
@@ -1,0 +1,2 @@
+test
+node
