+++ components/command/node/.ordering
@@ -1,0 +1,14 @@
+escape.mjs
+tokenize.mjs
+package.mjs
+mocha.mjs
+jest-config.mjs
+jest-argv.mjs
+jest.mjs
+node.mjs
+process.mjs
+remote.mjs
+node-recursive.mjs
+process-recursive.mjs
+remote-recursive.mjs
+index.mjs
