+++ components/command/node/escape.test.mjs
