+++ components/command/node/node.test.mjs
