+++ components/command/node/process-recursive.mjs
