+++ components/command/node/process-recursive.test.mjs
