+++ components/command/node/tokenize.test.mjs
